Update hashing dep graph to use gradle-tooling-based GradleProjectsProvider (#26)
This PR introduces a new implementation of the `GradleProjectsProvider`
interface that uses the gradle tooling api to fetch the set of active
projects for use in the hashing part of the system. The existing
implementation of the interface `SettingsGradleHashingProjectsProvider`
uses a settings.gradle-based approach, which is very fast, but the
primary expectation is that projects will be hashed only in CI as part
of the artifact publishing pipeline. This means the additional accuracy
from using the gradle apis directly is worth a slightly slower runtime.