Integrate picocli commands into Main.kt in :cli module (#24)
Updates the main class in the :cli module so that picocli commands are
run. Note, commented out remover command until #23 is merged.

Can be tested with:
```
./gradlew :cli:installShadowDist
./cli/build/install/cli-shadow/bin/cli hashing
```

This will fail with dependency graph issues, but does show that the
picocli part is running as we expert. Will update the dep graph with
fixes shortly