Update ArtifactSwapConfig to read values from gradle.properties file in current working directory, if present (#25)
We added this config class to encapsulate/centralize what data likely
needs to be provided by users in order for the system to work. The
original addition of this used static values for convenience. This PR
updates the implementation to check for the presence of
`gradle.properties` in the working directory where the tool is run, and
then loading relevant props from the file.

I tested this by building the cli tool, calling it from our internal
repo, and verifying the execution used the values in the props and
succeeded similar to how the internal tool behaves.