Correctly wire jvm args for gradle connection in task finder/runner code (#36)
Noticed in our CI executions that there were multiple gradle daemons
being created. This shouldn't be necessary, and can lead to excessive
memory usage. It was particularly surprising because the logs reported
that the daemon was started with a max memory value that was different
from what I passed on the command line. I traced the issue to a wiring
problem in koin where we are sometimes asking for "gradleJvmArgs" and
other times asking for/providing "jvmArgs".

This PR fixes the issue so that all the code requests jvm arguments for
gradle connections using the "jvmArgs" named identifier.