Go back to using -all as suffix for shadowJar task output (#79)
It was a nice idea to set the archive suffix to "", I like it
aesthetically, but it produces problems with gradle. Specifically,
setting "" as the suffix (e.g. `cli-0.1.2-SNAPSHOT.jar` instead of
`cli-0.1.2-all-SNAPSHOT.jar`) causes configuration problems for gradle
because the output from `:cli:shadowJar` task now appears to be an input
to other gradle tasks (this is because the output goes to the same
location as the other task, `:cli:jar`), but we don't declare
dependencies between `:cli:shadowJar` and those tasks. So, now I am
seeing gradle errors like this when running `./gradlew build`:
```
> Task :cli:shadowJar FAILED

[Incubating] Problems report is available at: file:///Users/wclausen/Development/worktrees/artifact-swap/wt2/build/reports/problems/problems-report.html

FAILURE: Build failed with an exception.

* What went wrong:
A problem was found with the configuration of task ':cli:shadowJar' (type 'ShadowJar').
  - Gradle detected a problem with the following location: '/Users/wclausen/Development/worktrees/artifact-swap/wt2/cli/build/libs/cli-0.1.2-SNAPSHOT.jar'.

    Reason: Task ':cli:generateJvmTestLintModel' uses this output of task ':cli:shadowJar' without declaring an explicit or implicit dependency. This can lead to incorrect results being produced, depending on what order the tasks are executed.

    Possible solutions:
      1. Declare task ':cli:shadowJar' as an input of ':cli:generateJvmTestLintModel'.
      2. Declare an explicit dependency on ':cli:shadowJar' from ':cli:generateJvmTestLintModel' using Task#dependsOn.
      3. Declare an explicit dependency on ':cli:shadowJar' from ':cli:generateJvmTestLintModel' using Task#mustRunAfter.

    For more information, please refer to https://docs.gradle.org/8.14.3/userguide/validation_problems.html#implicit_dependency in the Gradle documentation.

* Try:
> Declare task ':cli:shadowJar' as an input of ':cli:generateJvmTestLintModel'
> Declare an explicit dependency on ':cli:shadowJar' from ':cli:generateJvmTestLintModel' using Task#dependsOn
> Declare an explicit dependency on ':cli:shadowJar' from ':cli:generateJvmTestLintModel' using Task#mustRunAfter
```

This pr removes the configuration that causes the `shadowJar` task to
look like it depends on other tasks directly, when it does not/vice
versa.