Improve handling of android library publishing by considering flavors (#81)
Our existing plugin code assumes there are no defined product flavors
for android library modules that need to be published. This assumption
is not always true, and when it is violated, the plugins crash.

This PR updates the configuration of publishing to look for a _default_
product flavor and, if one exists, use the `debug` variant of that
flavor for publishing an aar/sources jar.

If flavors are defined but none are marked as "default", then the plugin
will continue to crash on configuration, as there is not enough
information to know which variant to publish.