Remove dangling 'named' usage from Koin configuration for Eventstream (#92)
In the refactoring to [move the downloader module out of
:core](https://github.com/block/artifact-swap/pull/86), some koin
configuration ended up getting out of sync. This PR fixes the di
configuration to provide an `Eventstream` object to classes that need
it. Simply removes the `named()` usage. There's only one `Eventstream`
object anyways, so no need for `named()` usage to distinguish between
multiple.