Add support for reading artifactory password from token file if password is not set with standard config value (#97)
We expect publishing artifacts from Ci to be a common workflow. It is
unlikely that CI secrets will be stored in a gradle.properties file, and
it is likely easier to set the artifactory password by reading from the
environment. This PR updates the password setting configuration to read
from the configured token file under the path specified by the
`SECRETS_PATH` env var.